\name{PGM-package}
\alias{PGM-package}
\alias{PGM}
\docType{package}
\title{
Poisson Graphical Models
}
\description{
	This packages includes multiple functions to implement the local Log-Linear Graphical Model based on pair-wise Poisson Markov Network using efficient, parallel algorithm named Poisson Graphical Lasso.
	This algorithm employs neighborhood selection to infer network structure. Stability selection is used in selecting the optimal network.
}
\details{
	\tabular{ll}{
	Package: \tab PGM\cr
	Type: \tab Package\cr
	Version: \tab Trial.2.0\cr
	Date: \tab 2014-03-04\cr
	License: \tab GPL2.0\cr
	}
	~~ An overview of how to use the package, including the most important ~~
	~~ functions ~~
}
\author{
	Who wrote it

	Maintainer: Who to complain to <yourfault@somewhere.net>
	~~ The author and/or maintainer of the package ~~
}
\references{
G.I. Allen and Z. Liu, 2012, A Log-Linear Graphical Model for Inferring Genetic Networks from High-Throughput Sequencing Data, \emph{The IEEE International Conference on Bioinformatics and Biomedicine (BIBM 2012)}.

E. Yang, P.K. Ravikumar, G.I. Allen, and Z. Liu, 2012, Graphical Models via Generalized Linear Models, \emph{NIPS}, \bold{vol. 25}, pp. 1367--1375.
}
\keyword{ package }
\seealso{
	~~ Optional links to other man pages, e.g. ~~
	~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
